Rename VeinHashPrivate -> DatabaseHash
Recent rename dance opened our eyes:

As the new name suggests StorageEventSystem is a Vein EventSystem. As such it
does adds much extra functionality e.g. get(Future)Component() which is not
business of an EventSystem exactly.

So start another refactoring and remove all this extra functionality and give
access to database which supports all these methods. The least hurting move
is rename and that is what this change is about.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>