START EXPERIMENTS ON INTROSPECTION: Transfer component data with introspection
One thing that avoids quick progress on vein clients: After receipt of
introspection it is upon client's duty to fetch all components.

So why not add component data to introspection to make this dance obsolete.
The data sent is not used yet so this change just causes extra payload and we
have to gain experience on target. If boot time regression is acceptable let's
start using initial component data in vf-qml, client storage and friends.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>